Finished moving the scheduling to init
