+++ dds_web/utils.py
@@ -15,6 +15,12 @@ from contextlib import contextmanager
 import flask
 import sqlalchemy
 
+# # imports related to scheduling
+import atexit
+
+import werkzeug
+from apscheduler.schedulers.background import BackgroundScheduler
+
 # Own modules
 from dds_web.database import models
 from dds_web import db, C_TZ
@@ -264,3 +270,111 @@ def permanent_delete():
     flask.current_app.logger.debug(
         "Permanently deleting the expired files (not implemented atm, just scheduled function)"
     )
+
+
+####################################################################################################
+# BACKGROUND SCHEDULER ###################################################### BACKGROUND SCHEDULER #
+####################################################################################################
+
+# Mind this discussion on the intricate peculiarities of running APScheduler in combination with Flask
+# https://github.com/agronholm/apscheduler/issues/250
+
+
+def scheduler_wrapper():
+
+    # Flask in debug mode spawns a child process so that it can restart the process each time the code changes,
+    # the new child process initializes and starts a new APScheduler, causing the jobs to be created twice
+    # within in the same database table:
+    # pymysql.err.IntegrityError: (1062, "Duplicate entry 'calc_costs' for key 'PRIMARY'") error
+
+    # Apparently, the reload is done with a subprocess.call, so we have 2 different Python interpreters running at the same time!
+    # This also means that any if statement or replace_existing=FALSE paramenter in add_job() won't prevent these errors.
+    # This if statement hopefully solves the issue:
+
+    if flask.helpers.get_debug_flag() and not werkzeug.serving.is_running_from_reloader():
+        return
+
+    scheduler = BackgroundScheduler(
+        {
+            "apscheduler.jobstores.default": {
+                "type": "sqlalchemy",
+                # "url": flask.current_app.config.get("SQLALCHEMY_DATABASE_URI"),
+                "engine": db.engine,
+            },
+            "apscheduler.timezone": "Europe/Stockholm",
+        }
+    )
+    flask.current_app.logger.info("Initiated main scheduler")
+
+    # Schedule invoicing calculations every 30 days
+    # TODO (ina): Change to correct interval - 30 days
+    if not scheduler.get_job("calc_costs"):
+        flask.current_app.logger.info("Added job: calc_costs")
+        scheduler.add_job(
+            invoice_units,
+            "cron",
+            id="calc_costs",
+            replace_existing=False,
+            coalesce=True,  # when several run times are due, none the less run the rob only once
+            month="1-12",
+            day="1-31",
+            hour="0",
+        )
+
+    # Schedule delete of rows in version table after a specific amount of time
+    # Currently: First of every month
+    if not scheduler.get_job("remove_versions"):
+        flask.current_app.logger.info("Added job: remove_versions")
+        scheduler.add_job(
+            remove_invoiced,
+            "cron",
+            id="remove_versions",
+            replace_existing=False,
+            coalesce=True,  # when several run times are due, none the less run the rob only once
+            month="1-12",
+            day="1",
+            hour="1",
+        )
+
+    # Schedule move of rows in files table after a specific amount of time
+    # to DeletedFiles (does not exist yet) table
+    # Currently: First of every month
+    if not scheduler.get_job("remove_expired"):
+        flask.current_app.logger.info("Added job: remove_expired")
+        scheduler.add_job(
+            remove_expired,
+            "cron",
+            id="remove_expired",
+            replace_existing=False,
+            coalesce=True,
+            month="1-12",
+            day="1",
+            hour="2",
+        )
+
+    # Schedule delete rows in expiredfiles table after a specific amount of time
+    # TODO (ina): Change interval - 1 day?
+    if not scheduler.get_job("permanent_delete"):
+        flask.current_app.logger.info("Added job: permanent_delete")
+        scheduler.add_job(
+            permanent_delete,
+            "cron",
+            id="permanent_delete",
+            replace_existing=False,
+            coalesce=True,
+            month="1-12",
+            day="1-31",
+            hour="3",
+        )
+
+    scheduler.start()
+    flask.current_app.logger.info("Started main scheduler")
+
+    # Shut down the scheduler when exiting the app
+    atexit.register(lambda: scheduler.shutdown())
+
+    # Print the currently scheduled jobs as verification:
+    joblist = scheduler.get_jobs()
+    flask.current_app.logger.info("Currently scheduled jobs:")
+    for job in joblist:
+        flask.current_app.logger.info(f"{job}")
